+++ demisto_sdk/commands/common/hook_validations/modeling_rule.py
@@ -1,0 +1,24 @@
+"""
+This module is designed to validate the correctness of generic definition entities in content.
+"""
+
+from demisto_sdk.commands.common.hook_validations.content_entity_validator import \
+    ContentEntityValidator
+
+
+class ModelingRuleValidator(ContentEntityValidator):
+    """
+    ModelingRuleValidator is designed to validate the correctness of the file structure we enter to content repo.
+    """
+
+    def is_valid_file(self, validate_rn=True, is_new_file=False, use_git=False):
+        """
+        Check whether the modeling rule is valid or not
+        """
+        return True
+
+    def is_valid_version(self):
+        """
+        May deleted or be edited in the future by the use of XSIAM new content
+        """
+        pass